修复无法进入多人模式的问题
