+++ Content/Items/HyacinthSeries/Projectile.ArethusaProj.cs
@@ -25,6 +25,10 @@ namespace Coralite.Content.Items.HyacinthSeries
 
         public ArethusaBullet()
         {
+            if (Main.dedServ)
+            {
+                return;
+            }
             Main.QueueMainThreadAction(() =>
             {
                 effect = new BasicEffect(Main.instance.GraphicsDevice);
