+++ Coralite.cs
@@ -5,14 +5,16 @@ global using Terraria.ModLoader;
 using Coralite.Compat.BossCheckList;
 using Coralite.Core;
 using Coralite.Core.Prefabs.Projectiles;
+using Coralite.Core.Systems.BossSystems;
 using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using Terraria;
 
 namespace Coralite
 {
-    public partial class Coralite : Mod
+    public class Coralite : Mod
     {
         public const int YujianHuluContainsMax = 3;
         public const int MaxParticleCount = 1201;
@@ -25,19 +27,41 @@ namespace Coralite
         public X2Smoother X2Smoother;
         public SinSmoother SinSmoother;
         public BezierEaseSmoother BezierEaseSmoother;
+
+        public Color RedJadeRed { get; private set; }
+        public Color IcicleCyan { get; private set; }
+        public Color MagicCrystalPink { get; private set; }
+        public Color CrystallineMagikePurple { get; private set; }
+        public Color SplendorMagicoreLightBlue { get; private set; }
+        public Color ThunderveinYellow { get; private set; }
         /// <summary>
         /// 从0快速接近1，之后快速返回0<br></br>
         /// 在0.5的时候到达1
         /// </summary>
         public ReverseX2Smoother ReverseX2Smoother;
 
-        //单例模式！
-        public static Coralite Instance { get; private set; }
+        private static Coralite _instance;
 
-        public Coralite()
+        //单例模式！
+        public static Coralite Instance
+        {
+            get
+            {
+                if (_instance == null)
+                {
+                    _instance = (Coralite)ModLoader.GetMod("Coralite");
+                }
+                return _instance;
+            }
+            set
+            {
+                _instance = value;
+            }
+        }
+
+        public override void Load()
         {
             Instance = this;
-
             NoSmootherInstance = new NoSmoother();
             HeavySmootherInstance = new HeavySmoother();
             SqrtSmoother = new SqrtSmoother();
@@ -46,10 +70,7 @@ namespace Coralite
             BezierEaseSmoother = new BezierEaseSmoother();
             ReverseX2Smoother = new ReverseX2Smoother();
             InitColor();
-        }
 
-        public override void Load()
-        {
             loadCache = new List<IOrderedLoadable>();
 
             foreach (Type type in Code.GetTypes())
@@ -66,13 +87,8 @@ namespace Coralite
             for (int k = 0; k < loadCache.Count; k++)
             {
                 loadCache[k].Load();
-                //SetLoadingText("Loading " + loadCache[k].GetType().Name);     //使用反射来显示加载的内容，但我暂时不需要（
+                SetLoadingText("Loading " + loadCache[k].GetType().Name);     //使用反射来显示加载的内容，但我暂时不需要（
             }
-
-            /////////////////////////////////其他乱七八糟的加载内容////////////////////////////////////////////
-
-            //由于代码写的比较优美所以没有在这里有其他乱七八糟的加载内容。
-
         }
 
         public override void Unload()
@@ -86,6 +102,8 @@ namespace Coralite
 
                 loadCache = null;
             }
+
+            Instance = null;
         }
 
         /// <summary>
@@ -104,5 +122,89 @@ namespace Coralite
             BossCheckListCalls.CallBossCheckList();
         }
 
+        public override object Call(params object[] args)
+        {
+            int argsLength = args.Length;
+            //Array.Resize(ref args, 15);
+
+            try
+            {
+                string methodName = args[0] as string;
+
+                switch (methodName)
+                {
+                    default: break;
+                    case "GetBossDowned"://获取击败BOSS的bool值
+                    case "BossDowned":
+                        if (argsLength < 2)
+                            return new ArgumentNullException("ERROR: Must specify a boss or event name as a string.");
+                        if (args[1] is not string)
+                            return new ArgumentException("ERROR: The argument to \"Downed\" must be a string.");
+                        return GetBossDowned(args[1].ToString());
+                }
+            }
+            catch (Exception e)
+            {
+                Logger.Error($"Call Error: {e.StackTrace} {e.Message}");
+            }
+
+            return new ArgumentException("ERROR: Invalid method name.");
+        }
+
+        /// <summary>
+        /// 获取击败BOSS的bool
+        /// </summary>
+        /// <param name="bossName"></param>
+        /// <returns></returns>
+        public static bool GetBossDowned(string bossName)
+        {
+            return bossName.ToLower() switch
+            {
+                "赤玉灵" or
+                "Rediancie"
+                    => DownedBossSystem.downedRediancie,
+                "冰龙宝宝" or
+                "BabyIceDragon"
+                    => DownedBossSystem.downedBabyIceDragon,
+                // "影子球" or "ShadowBalls" => DownedBossSystem.xxxx,
+                "荒雷龙" or
+                "ThunderveinDragon"
+                    => DownedBossSystem.downedThunderveinDragon,
+                "史莱姆皇帝" or
+                "至高帝史莱姆王" or
+                "至高帝·史莱姆王" or
+                "至高帝" or
+                "SlimeEmperor"
+                    => DownedBossSystem.downedSlimeEmperor,
+                "赤血玉灵" or
+                "血咒精赤玉灵" or
+                "血咒精·赤玉灵" or
+                "血咒精" or
+                "Bloodiancie"
+                    => DownedBossSystem.downedBloodiancie,
+                "梦魇之花" or
+                "梦界主世纪之花" or
+                "梦界主·世纪之花" or
+                "梦界主" or
+                "NightmarePlantera" or
+                "Nightmare Plantera"
+                    => DownedBossSystem.downedNightmarePlantera,
+                _ => false,
+            };
+        }
+
+        public void InitColor()
+        {
+            if (Main.dedServ)
+            {
+                return;
+            }
+            RedJadeRed = new Color(221, 50, 50);
+            IcicleCyan = new Color(43, 255, 198);
+            MagicCrystalPink = new Color(255, 190, 236);
+            CrystallineMagikePurple = new Color(140, 130, 252);
+            SplendorMagicoreLightBlue = new Color(190, 225, 235);
+            ThunderveinYellow = new Color(255, 202, 101);
+        }
     }
 }
