+++ Localization/en-US/Mods.Coralite.Projectiles.hjson
@@ -533,3 +533,4 @@ ArcaneFlameDaggerProj.DisplayName: Arcane Flame Dagger Proj
 ArcaneFlameDaggerSlash.DisplayName: Arcane Flame Dagger Slash
 MagikeWorldBallProj.DisplayName: Magike World Ball Proj
 MagicCrystalBobber.DisplayName: Magic Crystal Bobber
+FilterProj.DisplayName: Filter Proj
