+++ Localization/zh-Hans/Mods.Coralite.Projectiles.hjson
@@ -533,3 +533,4 @@ TerranascenceSpurt.DisplayName: 泰拉勇者的剑
 // ArcaneFlameDaggerSlash.DisplayName: Arcane Flame Dagger Slash
 // MagikeWorldBallProj.DisplayName: Magike World Ball Proj
 // MagicCrystalBobber.DisplayName: Magic Crystal Bobber
+// FilterProj.DisplayName: Filter Proj
